+++ system/src/Grav/Common/Uri.php
@@ -482,12 +482,19 @@ class Uri
      *
      * @return string the more friendly formatted url
      */
-    public static function convertUrl(Page $page, $markdown_url)
+    public static function convertUrl(Page $page, $markdown_url, $type = 'link')
     {
         $grav = Grav::instance();
 
+        // Link processing should prepend language
+        $language_append = '';
+        if ($type == 'link') {
+            $active_language = $grav['language']->getActive();
+            $language_append = $active_language ? '/'.$active_language : '';
+        }
+
         $pages_dir = $grav['locator']->findResource('page://');
-        $base_url = rtrim($grav['base_url'] . $grav['pages']->base(), '/');
+        $base_url = rtrim($grav['base_url'] . $grav['pages']->base(), '/') . $language_append;
 
         // if absolute and starts with a base_url move on
         if (pathinfo($markdown_url, PATHINFO_DIRNAME) == '.' && $page->url() == '/') {
@@ -548,7 +555,7 @@ class Uri
             $instances = $grav['pages']->instances();
             if (isset($instances[$page_path])) {
                 $target = $instances[$page_path];
-                $url_bits['path'] = $base_url . $target->route() . $filename;
+                $url_bits['path'] = $base_url . rtrim($target->route(), '/') . $filename;
                 return Uri::buildUrl($url_bits);
             }