+++ system/src/Grav/Console/ConsoleTrait.php
@@ -107,4 +107,21 @@ trait ConsoleTrait
         $input = new ArrayInput($all);
         return $command->run($input, $this->output);
     }
+    
+    /**
+     * Validate if the system is based on windows or not.
+     * 
+     * @return bool
+     */
+    public function isWindows()
+    {
+        $keys = [
+            'CYGWIN_NT-5.1',
+            'WIN32',
+            'WINNT',
+            'Windows'
+        ];
+
+        return array_key_exists(PHP_OS, $keys);
+    }
 }